health_history_service/requirements.txt
flask==2.0.1
werkzeug==2.0.3
mysql-connector-python==8.0.26
python-dotenv==0.19.0 